Address UX feedback (batch 2): split Configure, Discover bulk keybinds, Done weighting
Three items in one commit:

1. Split Configure into two screens. The previous single screen had
   six radio panels stacked, which was the worst density offender in
   the flow. Split along the natural cut:
   - 6 Configure (workflow): release strategy / credentials /
     augmentation — "when does this workflow run, how does it
     authenticate, how does it source metadata?"
   - 7 Configure (SBOM): enrichment / formats / build provenance —
     "what goes into each SBOM, do we sign it?"
   TOTAL_STEPS bumped 7 → 8; Review / Apply / Done bump step_index
   to 8; Welcome's 'What we'll do' preview list lengthens. Old
   configure.py is removed; the navigation goes Components →
   ConfigureWorkflowScreen → ConfigureSbomScreen → ReviewScreen.

2. Discover bulk-selection keybinds. With 20 lockfiles tab-and-space
   gets tedious; added 'a' to select all, 'n' to select none, both
   shown in the footer. The silent app.bell() on an empty submit
   now also surfaces a visible status line ("Pick at least one
   lockfile to continue — Space to toggle, a to select all").

3. Done weighting. The applied summary used the same green ✓ glyph
   for every component regardless of whether it was created or
   reused. Re-running the wizard against an already-onboarded
   workspace shouldn't make every line feel like an event. Reused
   components now get a muted '·' marker and gray text; newly
   created ones keep the green '+' + bolded id + 'created' tag.
   Cross-references state.component_ids against
   plan.create_components to figure out which is which.
